+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 * add ability to promote gems in the order of being updated
 * add ability to demote gems in the order of being updated
 * add ability to only update specific gems
+* add ability to update everything but specific excluded gems
 
 ## 0.0.1 (2018-08-01)